tests: Swapping unit tests cleanup (#5226)
* refactor: cleanup DCA tests

* chore: split tests into files

* refactor: cleanup swapping tests

* test: add zero chunk size test

* refactor: address PR comments

* test: add missing safe mode test

* feat: add then_process_blocks_until_block utility

* chore: avoid INIT_BLOCK sanity check

---------
