+++ state-chain/pallets/cf-swapping/src/tests/fill_or_kill.rs
@@ -3,11 +3,11 @@ use super::*;
 const BROKER_FEE: AssetAmount = INPUT_AMOUNT * BROKER_FEE_BPS as u128 / 10_000;
 
 fn fok_swap(refund_params: Option<TestRefundParams>) -> TestSwapParams {
-	params(None, refund_params, false)
+	TestSwapParams::new(None, refund_params, false)
 }
 
 fn fok_swap_ccm(refund_params: Option<TestRefundParams>) -> TestSwapParams {
-	params(None, refund_params, true)
+	TestSwapParams::new(None, refund_params, true)
 }
 
 #[track_caller]
@@ -31,9 +31,7 @@ fn both_fok_and_regular_swaps_succeed_first_try() {
 	const FOK_REQUEST_ID: u64 = 2;
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			const REFUND_PARAMS: TestRefundParams = TestRefundParams {
 				retry_duration: DEFAULT_SWAP_RETRY_DELAY_BLOCKS,
 				min_output: (INPUT_AMOUNT - BROKER_FEE) * DEFAULT_SWAP_RATE,
@@ -60,7 +58,7 @@ fn both_fok_and_regular_swaps_succeed_first_try() {
 				SWAPS_SCHEDULED_FOR_BLOCK,
 			);
 		})
-		.then_execute_at_block(SWAPS_SCHEDULED_FOR_BLOCK, |_| {})
+		.then_process_blocks_until_block(SWAPS_SCHEDULED_FOR_BLOCK)
 		.then_execute_with(|_| {
 			assert_event_sequence!(
 				Test,
@@ -100,9 +98,7 @@ fn price_limit_is_respected_in_fok_swap() {
 	const FOK_SWAP_2_ID: u64 = 3;
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			insert_swaps(&vec![
 				fok_swap(None),
 				fok_swap(Some(TestRefundParams {
@@ -120,7 +116,7 @@ fn price_limit_is_respected_in_fok_swap() {
 				SWAPS_SCHEDULED_FOR_BLOCK,
 			);
 		})
-		.then_execute_at_block(3u64, |_| {})
+		.then_process_blocks_until_block(3u64)
 		.then_execute_with(|_| {
 			assert_eq!(System::block_number(), SWAPS_SCHEDULED_FOR_BLOCK);
 			// Swap 2 should fail due to price limit and rescheduled for block
@@ -171,9 +167,7 @@ fn fok_swap_gets_refunded_due_to_price_limit() {
 		SWAPS_SCHEDULED_FOR_BLOCK + (DEFAULT_SWAP_RETRY_DELAY_BLOCKS as u64);
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			// Min output for swap 1 is too high to be executed:
 			const MIN_OUTPUT: AssetAmount = (INPUT_AMOUNT - BROKER_FEE) * DEFAULT_SWAP_RATE + 1;
 			insert_swaps(&[fok_swap(Some(TestRefundParams {
@@ -188,7 +182,7 @@ fn fok_swap_gets_refunded_due_to_price_limit() {
 				SWAPS_SCHEDULED_FOR_BLOCK,
 			);
 		})
-		.then_execute_at_block(SWAPS_SCHEDULED_FOR_BLOCK, |_| {})
+		.then_process_blocks_until_block(SWAPS_SCHEDULED_FOR_BLOCK)
 		.then_execute_with(|_| {
 			// Swap 1 should fail here and rescheduled for a later block,
 			// but swap 2 (without FoK parameters) should still be successful:
@@ -208,7 +202,7 @@ fn fok_swap_gets_refunded_due_to_price_limit() {
 				}),
 			);
 		})
-		.then_execute_at_block(SWAP_RETRIED_AT_BLOCK, |_| {})
+		.then_process_blocks_until_block(SWAP_RETRIED_AT_BLOCK)
 		.then_execute_with(|_| {
 			// Swap request should be removed in case of refund
 			assert_eq!(SwapRequests::<Test>::get(FOK_SWAP_REQUEST_ID), None);
@@ -238,9 +232,7 @@ fn fok_swap_gets_refunded_due_to_price_impact_protection() {
 	const REGULAR_SWAP_ID: u64 = 2;
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			// FoK swap 1 should fail and will eventually be refunded
 			insert_swaps(&[fok_swap(Some(TestRefundParams {
 				retry_duration: DEFAULT_SWAP_RETRY_DELAY_BLOCKS,
@@ -274,7 +266,7 @@ fn fok_swap_gets_refunded_due_to_price_impact_protection() {
 				})
 			);
 		})
-		.then_execute_at_block(SWAP_RETRIED_AT_BLOCK, |_| {})
+		.then_process_blocks_until_block(SWAP_RETRIED_AT_BLOCK)
 		.then_execute_with(|_| {
 			// Swap request should be removed in case of refund
 			assert_eq!(SwapRequests::<Test>::get(FOK_SWAP_REQUEST_ID), None);
@@ -301,7 +293,7 @@ fn fok_test_zero_refund_duration() {
 	const SWAPS_SCHEDULED_FOR_BLOCK: u64 = INIT_BLOCK + SWAP_DELAY_BLOCKS as u64;
 
 	new_test_ext()
-		.execute_with(|| {
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			// A swap with 0 retry duration should be tried exactly 1 time
 			insert_swaps(&[fok_swap(Some(TestRefundParams {
 				retry_duration: 0,
@@ -337,9 +329,7 @@ fn fok_ccm_happy_path() {
 	const EXPECTED_OUTPUT: AssetAmount = (INPUT_AMOUNT - BROKER_FEE) * DEFAULT_SWAP_RATE;
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			insert_swaps(&[fok_swap_ccm(Some(TestRefundParams {
 				retry_duration: DEFAULT_SWAP_RETRY_DELAY_BLOCKS,
 				min_output: EXPECTED_OUTPUT,
@@ -347,7 +337,7 @@ fn fok_ccm_happy_path() {
 
 			assert_swaps_scheduled_for_block(&[PRINCIPAL_SWAP_ID], PRINCIPAL_BLOCK);
 		})
-		.then_execute_at_block(PRINCIPAL_BLOCK, |_| {})
+		.then_process_blocks_until_block(PRINCIPAL_BLOCK)
 		.then_execute_with(|_| {
 			assert_event_sequence!(
 				Test,
@@ -359,7 +349,7 @@ fn fok_ccm_happy_path() {
 				}),
 			);
 		})
-		.then_execute_at_block(GAS_BLOCK, |_| {})
+		.then_process_blocks_until_block(GAS_BLOCK)
 		.then_execute_with(|_| {
 			assert_event_sequence!(
 				Test,
@@ -383,9 +373,7 @@ fn fok_ccm_refunded() {
 	const PRINCIPAL_AMOUNT: AssetAmount = INPUT_AMOUNT - GAS_BUDGET;
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			insert_swaps(&[fok_swap_ccm(Some(TestRefundParams {
 				retry_duration: 0,
 				min_output: INPUT_AMOUNT * DEFAULT_SWAP_RATE + 1,
@@ -403,7 +391,7 @@ fn fok_ccm_refunded() {
 
 			assert_swaps_scheduled_for_block(&[PRINCIPAL_SWAP_ID], PRINCIPAL_BLOCK);
 		})
-		.then_execute_at_block(PRINCIPAL_BLOCK, |_| {})
+		.then_process_blocks_until_block(PRINCIPAL_BLOCK)
 		.then_execute_with(|_| {
 			assert_event_sequence!(
 				Test,
@@ -431,9 +419,7 @@ fn fok_ccm_refunded_no_gas_swap() {
 	const OUTPUT_ASSET: Asset = Asset::Usdc;
 
 	new_test_ext()
-		.execute_with(|| {
-			assert_eq!(System::block_number(), INIT_BLOCK);
-
+		.then_execute_at_block(INIT_BLOCK, |_| {
 			let refund_params = TestRefundParams {
 				retry_duration: 0,
 				min_output: INPUT_AMOUNT * DEFAULT_SWAP_RATE + 1,
@@ -462,7 +448,7 @@ fn fok_ccm_refunded_no_gas_swap() {
 
 			assert_swaps_scheduled_for_block(&[PRINCIPAL_SWAP_ID], PRINCIPAL_BLOCK);
 		})
-		.then_execute_at_block(PRINCIPAL_BLOCK, |_| {})
+		.then_process_blocks_until_block(PRINCIPAL_BLOCK)
 		.then_execute_with(|_| {
 			assert_event_sequence!(
 				Test,
